rcws_project/requirements.txt
# Django 및 핵심 패키지
Django==4.2.7
djangorestframework==3.15.2
django-cors-headers==4.3.1

# 데이터베이스
psycopg2-binary==2.9.9

# 실시간 통신
channels==4.2.2
channels-redis==4.1.0

# 비동기 작업
celery==5.3.4
redis==5.0.1

# API 문서화
drf-yasg==1.21.10

# 이미지 처리
Pillow==10.1.0

# 환경 변수 관리
python-decouple==3.8

# 로깅 및 모니터링
sentry-sdk==1.38.0

# 보안
django-ratelimit==4.1.0

# 파일 업로드
django-storages==1.14.2

# 이메일 템플릿
django-templated-mail==1.1.1

# 개발 도구 (개발 환경에서만)
# django-debug-toolbar==4.2.0
# django-extensions==3.2.3 